tonnes of highlight stuff

+++ src/admin/highlights/ApproveHighlights.jsx
@@ -6,11 +6,10 @@ import AddIcon from '@material-ui/icons/Add';
 import defaultStyles from './ApproveHighlights.module.scss';
 import {
     fetchHighlightsForApprovalRequest, approveHighlightRequest, rejectHighlightRequest,
-    deleteHighlightRequest, fetchAllRejectedHighlightsRequest
+    deleteHighlightRequest, fetchAllRejectedHighlightsRequest, reapproveRejectedHighlightRequest
 } from '../actions';
 import YouTubeList from '../../common/youtubelist/YouTubeList';
 import StyledButton from '../../common/StyledButton/StyledButton';
-import ConfirmModal from '../../common/modal/ConfirmModal';
 import StyledModal from '../../common/modal/StyledModal';
 import StyledInput from '../../common/StyledInput/StyledInput';
 import Spinner from '../../common/spinner/Spinner';
@@ -24,35 +23,67 @@ const columns = [
     {
         id: 'title',
         label: 'Title',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: true
     },
     {
         id: 'author',
         label: 'Author',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: true
     },
     {
         id: 'videoLink',
         label: 'Video',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: true
     },
     {
         id: 'dateCreated',
         label: 'Date Created',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: true
+    },
+    {
+        id: 'rejectedBy',
+        label: 'Rejected By',
+        align: 'center',
+        approved: false,
+        rejected: true
     },
     {
         id: 'upvotes',
         label: 'Upvotes',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: false
+    },
+    {
+        id: 'reason',
+        label: 'Reason',
+        align: 'center',
+        approved: false,
+        rejected: true
     },
     {
         id: 'delete',
         label: '',
-        align: 'center'
+        align: 'center',
+        approved: true,
+        rejected: true
     }
 ];
 
+const modalOptions = {
+    REAPPROVE: 'REAPPROVE',
+    DELETE: 'DELETE',
+    REJECT: 'REJECT',
+    APPROVE: 'APPROVE'
+};
 
 const ApproveHighlights = props => {
     useEffect(() => {
@@ -63,49 +94,57 @@ const ApproveHighlights = props => {
         props.fetchAllRejectedHighlightsRequest]);
 
     const [activeHightlight, setActiveHighlight] = useState('');
-    const [confirmModalOpen, setConfirmModalOpen] = useState(false);
-    const [rejectModalOpen, setRejectModalOpen] = useState(false);
     const [reasonToReject, setReasonToReject] = useState('');
-    const [deleteModalOpen, setDeleteModalOpen] = useState(false);
     const [filterBy, setFilterBy] = useState('allTime');
     const [searchBy, setSearchBy] = useState('');
 
-    const openConfirm = useCallback(id => {
-        setActiveHighlight(id);
-        setConfirmModalOpen(true);
-    }, []);
+    const [filterByRejected, setFilterByRejected] = useState('allTime');
+    const [searchByRejected, setSearchByRejected] = useState('');
 
-    const openReject = useCallback(id => {
-        setActiveHighlight(id);
-        setRejectModalOpen(true);
-    }, []);
+    const [modalAction, setModalAction] = useState('');
+    const [fancyModalOpen, setFancyModalOpen] = useState(false);
+    const [modalText, setModalText] = useState('');
 
-    const openDelete = useCallback(id => {
-        setActiveHighlight(id);
-        setDeleteModalOpen(true);
-    }, []);
-
-    const closeModal = useCallback(() => {
-        setConfirmModalOpen(false);
-        setRejectModalOpen(false);
-        setDeleteModalOpen(false);
+    const closeFancyModal = useCallback(() => {
+        setFancyModalOpen(false);
+        setModalAction('');
         setActiveHighlight('');
+    }, [fancyModalOpen, setFancyModalOpen, modalAction,
+        setModalAction, activeHightlight, setActiveHighlight]);
+
+    const confirmFancyModal = useCallback(() => {
+        if (modalAction === modalOptions.DELETE) {
+            props.deleteHighlightRequest(activeHightlight, reasonToReject);
+        }
+        if (modalAction === modalOptions.APPROVE) {
+            props.approveHighlightRequest(activeHightlight);
+        }
+        if (modalAction === modalOptions.REJECT) {
+            props.rejectHighlightRequest(activeHightlight, reasonToReject);
+        }
+        if (modalAction === modalOptions.REAPPROVE) {
+            props.reapproveRejectedHighlightRequest(activeHightlight);
+        }
+        closeFancyModal();
+    }, [modalAction, activeHightlight, fancyModalOpen]);
+
+    const openFancyModal = useCallback((id, action) => {
+        if (action === modalOptions.REAPPROVE) {
+            setModalText('Reapprove highlight');
+        }
+        if (action === modalOptions.DELETE) {
+            setModalText('Delete highlight');
+        }
+        if (action === modalOptions.APPROVE) {
+            setModalText('Approve highlight');
+        }
+        if (action === modalOptions.REJECT) {
+            setModalText('Reject Highlight');
+        }
+        setActiveHighlight(id);
+        setModalAction(action);
+        setFancyModalOpen(true);
     }, []);
-
-    const confirm = useCallback(() => {
-        props.approveHighlightRequest(activeHightlight);
-        closeModal();
-    }, [props.approveHighlightRequest, activeHightlight]);
-
-    const reject = useCallback(() => {
-        props.rejectHighlightRequest(activeHightlight, reasonToReject);
-        closeModal();
-    }, [props.approveHighlightRequest, activeHightlight, reasonToReject]);
-
-    const deleteConfirm = useCallback(() => {
-        props.deleteHighlightRequest(activeHightlight, reasonToReject);
-        closeModal();
-    }, [props.deleteHighlightRequest, activeHightlight, reasonToReject]);
 
     const mapRows = (videos, deleteSymbol) => videos.map(x => ({
         id: x.id,
@@ -114,22 +153,29 @@ const ApproveHighlights = props => {
         videoLink: generateYouTubeLinkFromId(x.videoId),
         dateCreated: generateTimeSinceNow(x.dateCreated),
         upvotes: `${x.upvotes.length - x.downvotes.length > 0 ? '+' : ''}${x.upvotes.length - x.downvotes.length}`,
+        reason: x.reason,
+        rejectedBy: x.rejectedBy,
         delete:
     <div className={props.styles.deleteIcon}>
-        {deleteSymbol ? <DeleteIcon color="primary" onClick={() => openDelete(x.id)} /> : <AddIcon color="secondary" />}
+        {deleteSymbol ? <DeleteIcon color="primary" onClick={() => openFancyModal(x.id, modalOptions.DELETE)} />
+            : <AddIcon color="secondary" onClick={() => openFancyModal(x.id, modalOptions.REAPPROVE)} />}
     </div>
     }));
 
     return (
         <>
             <div className={props.styles.approveHighlightsWrapper}>
-                Here you can approve / reject highlights. Please give a reason when rejecting
+                <div className={props.styles.highlightInfo}>
+                Here you can approve / reject highlights. Please give a reason when rejecting.
+                You can also delete active highlights and reapprove rejected ones.
+                Rejected highlights are automatically once they are at least a month old.
+                </div>
             </div>
             <div className={props.styles.highlightsWrapper}>
                 <YouTubeList
                     approversPage
-                    openConfirm={openConfirm}
-                    openReject={openReject}
+                    openConfirm={id => openFancyModal(id, modalOptions.APPROVE)}
+                    openReject={id => openFancyModal(id, modalOptions.REJECT)}
                     rejectHighlightRequest={props.rejectHighlightRequest}
                     videos={props.highlightsForApproval}
                 />
@@ -158,64 +204,56 @@ const ApproveHighlights = props => {
                         }))}
                         value={filterBy}
                     />
-
                 </div>
                 <div className={props.styles.videoSearchFilter}>
                     <StyledInput label="Filter by author / title" onChange={setSearchBy} value={searchBy} />
                 </div>
                 <div className={props.styles.gridWrapper}>
                     <Grid
-                        gridHeader="All Highlights"
-                        columns={columns}
+                        gridHeader="Approved Highlights"
+                        columns={columns.filter(x => x.approved)}
                         rows={mapRows(helpers.filterByDate(filterBy, props.videos, searchBy), true)}
                     />
                 </div>
             </div>
 
             <div className={props.styles.allRejectedVideos}>
+                <div className={props.styles.dateFilters}>
+                    <RadioButton
+                        radioLabel="Filter By Date"
+                        onChange={setFilterByRejected}
+                        options={Object.values(helpers.dateFilters).map(x => ({
+                            radioLabel: x.label,
+                            value: x.id
+                        }))}
+                        value={filterByRejected}
+                    />
+                </div>
+                <div className={props.styles.videoSearchFilter}>
+                    <StyledInput label="Filter by author / title" onChange={setSearchByRejected} value={searchByRejected} />
+                </div>
                 <div className={props.styles.gridWrapper}>
                     <Grid
-                        gridHeader="Rejected"
-                        columns={columns}
-                        rows={mapRows(helpers.filterByDate(filterBy, props.rejectedHighlights, searchBy), false)}
+                        gridHeader="Rejected Highlights"
+                        columns={columns.filter(x => x.rejected)}
+                        rows={mapRows(helpers.filterByDate(filterByRejected,
+                            props.rejectedHighlights, searchByRejected), false)}
                     />
                 </div>
             </div>
-
-            <ConfirmModal
-                cancel={closeModal}
-                closeModal={closeModal}
-                isOpen={confirmModalOpen}
-                submit={confirm}
-                text="Are you sure you want to approve this video?"
-            />
             <StyledModal
                 backdrop
-                closeModal={closeModal}
+                closeModal={closeFancyModal}
                 error
-                isOpen={rejectModalOpen}
-                headerMessage="Reject Video"
+                isOpen={fancyModalOpen}
+                headerMessage={modalText}
             >
                 <div className={props.styles.modalWrapper}>
-                    <div><StyledInput label="Reason for rejection" onChange={setReasonToReject} value={reasonToReject} /></div>
+                    {(modalAction === modalOptions.REJECT || modalAction === modalOptions.DELETE)
+                    && <div><StyledInput label="Reason" onChange={setReasonToReject} value={reasonToReject} /></div>}
                     <div className={props.styles.modalButtons}>
-                        <StyledButton text="Confirm" onClick={reject} />
-                        <StyledButton text="Cancel" color="secondary" onClick={closeModal} />
-                    </div>
-                </div>
-            </StyledModal>
-            <StyledModal
-                backdrop
-                closeModal={closeModal}
-                error
-                isOpen={deleteModalOpen}
-                headerMessage="Delete Video"
-            >
-                <div className={props.styles.modalWrapper}>
-                    <div><StyledInput label="Reason for deletion" onChange={setReasonToReject} value={reasonToReject} /></div>
-                    <div className={props.styles.modalButtons}>
-                        <StyledButton text="Confirm" onClick={deleteConfirm} />
-                        <StyledButton text="Cancel" color="secondary" onClick={closeModal} />
+                        <StyledButton text="Confirm" onClick={confirmFancyModal} />
+                        <StyledButton text="Cancel" color="secondary" onClick={closeFancyModal} />
                     </div>
                 </div>
             </StyledModal>
@@ -243,6 +281,7 @@ ApproveHighlights.propTypes = {
         id: PropTypes.string
     })),
     loadingHighlightsForApproval: PropTypes.bool,
+    reapproveRejectedHighlightRequest: PropTypes.func.isRequired,
     rejectedHighlights: PropTypes.arrayOf(PropTypes.shape({})),
     rejectHighlightRequest: PropTypes.func.isRequired,
     styles: PropTypes.objectOf(PropTypes.string),
@@ -255,6 +294,7 @@ const mapDispatchToProps = {
     fetchAllRejectedHighlightsRequest,
     fetchHighlightsRequest,
     fetchHighlightsForApprovalRequest,
+    reapproveRejectedHighlightRequest,
     rejectHighlightRequest
 };
